Add JsonExceptionListener and register it in services
A new JsonExceptionListener is added to handle exceptions and give responses in JSON format. It is registered as a service and tagged to kernel.exception event in services.yaml, allowing to provide exception details in a structured and standard format. This improves debugging especially in the 'dev' environment; where more details, including the request info, are provided.